Refer to unity8 for installation instructions
I notice that many are reporting issues with Unity8 in the wrong issue tracker, namely on the Unity8 installer issue tracker, probably because that's where the installation instructions are found.   Removing the installation instructions from this readme (and pointing this readme to the readme found in the Unity8 repository) may reduce this problem moving forward.